+++ movierecommendationsystem-frontend/src/app/tv-shows/tv-show-details/tv-show-details.component.ts
@@ -10,6 +10,10 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
 import { TrailerDialogComponent } from '../../trailer-dialog/trailer-dialog.component';
 import { FormsModule } from '@angular/forms';
+import { TvShow } from '../../model/tvShow';
+import { Interaction } from '../../model/interaction';
+import { User } from '../../model/user';
+import { MediaVideo } from '../../model/mediaVideo';
 
 @Component({
   selector: 'app-tv-show-details',
@@ -19,15 +23,15 @@ import { FormsModule } from '@angular/forms';
   styleUrl: './tv-show-details.component.css'
 })
 export class TvShowDetailsComponent implements OnInit {
-  tvShow : any;
+  tvShow : TvShow | null = null;
   isWatchList : boolean = false;
   isFavorite : boolean = false;
   username : string = '';
-  user : any;
+  user : User | null = null;
   tvShowId : number = 0;
-  interaction : any;
+  interaction : Interaction | null = null;
   userRating : number = 0;
-  trailer : any;
+  trailer : MediaVideo | null = null;
   constructor(private authService : AuthService,private tvShowService : TvShowService,
     private userService : UserService, private interactionService : InteractionService,
     private route : ActivatedRoute, private router : Router,
@@ -41,14 +45,18 @@ export class TvShowDetailsComponent implements OnInit {
       response => {this.username = response.username;
         this.userService.findByUsename(this.username).subscribe(
           response => {this.user = response;
-            this.interactionService.findInteraction(this.user.userId,this.tvShowId).subscribe(
-              response => {this.interaction = response;
-                this.userRating = this.interaction.rating;
-                this.isWatchList = this.interaction.watchList;
-                this.isFavorite = this.interaction.favorite;
-              },
-              error => {console.error('error fetching interaction',error);}
-            );
+            if(this.user){
+              this.interactionService.findInteraction(this.user.userId,this.tvShowId).subscribe(
+                response => {this.interaction = response;
+                  if(this.interaction){
+                    this.userRating = this.interaction.rating;
+                    this.isWatchList = this.interaction.watchList;
+                    this.isFavorite = this.interaction.favorite;
+                  }       
+                },
+                error => {console.error('error fetching interaction',error);}
+              );
+            }
           }
         );
       },
@@ -71,7 +79,7 @@ export class TvShowDetailsComponent implements OnInit {
   }
 
   addToWatchList(){
-    if(this.username){
+    if(this.username && this.user){
       this.interactionService.addToWatchList(this.user.userId,this.tvShowId,'tv show').subscribe(
         response => {console.log('added with success',response);
         },
@@ -84,7 +92,7 @@ export class TvShowDetailsComponent implements OnInit {
   }
 
   removeFromWatchList(){
-    if(this.username){
+    if(this.username && this.user){
       this.interactionService.removeFromWatchList(this.user.userId,this.tvShowId).subscribe(
         response => {console.log('removed with success',response);
         },
@@ -97,7 +105,7 @@ export class TvShowDetailsComponent implements OnInit {
   }
 
   addToFavorite(){
-    if(this.username){
+    if(this.username && this.user){
       this.interactionService.addToFavorite(this.user.userId,this.tvShowId,'tv show').subscribe(
         response => {console.log('added with success',response);
         },
@@ -110,7 +118,7 @@ export class TvShowDetailsComponent implements OnInit {
   }
 
   removeFromFavorite(){
-    if(this.username){
+    if(this.username && this.user){
       this.interactionService.removeFromFavorite(this.user.userId,this.tvShowId).subscribe(
         response => {console.log('removed with success',response);
         },
@@ -149,13 +157,15 @@ export class TvShowDetailsComponent implements OnInit {
   }
 
   rateTvShow(){
-    this.interactionService.rate(this.user.userId,this.tvShowId,this.userRating).subscribe(
-      response => {this.ngOnInit()},
-      error => {console.error('error rating the movie',error);}
-    );
+    if(this.username && this.user){
+      this.interactionService.rate(this.user.userId,this.tvShowId,this.userRating).subscribe(
+        response => {this.ngOnInit()},
+        error => {console.error('error rating the movie',error);}
+      );
+    }
   }
 
   seasonDetails(seasonNumber : number){
-    this.router.navigate(['/tv-show',this.tvShowId,'season',seasonNumber]);
+    this.router.navigate(['/tv-shows',this.tvShowId,'seasons',seasonNumber]);
   }
 }
